Restructure gateway interfaces and implementations Add summary blocks for gateways
