--- src/IL2CareerModel/Data/Gateways/CareerGateway.cs
@@ -1,34 +1,0 @@
-﻿using IL2CarrerReviverModel.Data.Repositories;
-using IL2CarrerReviverModel.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace IL2CarrerReviverModel.Data.Gateways;
-internal sealed class CareerGateway : ICareerGateway
-{
-    private readonly IBaseRepository<Career, long> repository;
-
-    public CareerGateway(IBaseRepository<Career, long> repository)
-    {
-        this.repository = repository;
-    }
-
-    public Career? GetById(long id)
-    {
-        return repository.GetById(id);
-    }
-
-    public IEnumerable<Career> GetAll()
-    {
-        return repository.GetAll();
-    }
-
-    public Career? Update(Career career)
-    {
-        repository.Update(career);
-        return GetById(career.Id);
-    }
-}
