+++ src/IL2CareerModel/Data/Gateways/ISortieGateway.cs
@@ -1,16 +1,12 @@
-﻿using IL2CarrerReviverModel.Models;
+﻿using IL2CareerModel.Data.Gateways;
+using IL2CarrerReviverModel.Models;
 
 namespace IL2CarrerReviverModel.Data.Gateways;
-public interface ISortieGateway
+
+/// <summary>
+/// The gateway to change the sortie table on the database
+/// </summary>
+public interface ISortieGateway : IReadGateway<Sortie>, IWriteGateway<Sortie>
 {
-    IEnumerable<Sortie> GetAll() => GetAll(_ => true);
-
-    IEnumerable<Sortie> GetAll(Func<Sortie, bool> filter);
-
-    Sortie? GetById(long id);
-
-    bool DeleteById(long id);
-
-    bool Delete(Sortie sortie);
 }
