+++ src/IL2CareerModel/DependencyInjection/ModelDependenciesExtension.cs
@@ -1,4 +1,5 @@
-﻿using IL2CarrerReviverConsole.Services;
+﻿using IL2CareerModel.Data.Gateways.Implementations;
+using IL2CarrerReviverConsole.Services;
 using IL2CarrerReviverModel.Data;
 using IL2CarrerReviverModel.Data.Gateways;
 using IL2CarrerReviverModel.Data.Repositories;
